+++ packages/streams/src/methods/verifySignature.ts
